+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Scraping of tenant cluster prometheus
 - Scraping of control plane prometheus
 - Add installation label
+- Add labelling schema alert
 
 ### Changed